requirements.txt
streamlit==1.8.0
pyyaml==6.0
pyperclip==1.8.2